refactor: extract widget settings panels to separate components
